+++ BackEndFlask/models/course.py
@@ -1,23 +1,11 @@
 from core import db
-from flask_login import UserMixin
-from sqlalchemy import ForeignKey, Enum, Boolean
 from sqlalchemy.exc import SQLAlchemyError
+from models.schemas import Course
 
 class InvalidCourseID(Exception):
     "Raised when course_id does not exist!!!"
     pass
 
-class Course(UserMixin, db.Model):
-    __tablename__ = "Course"
-    __table_args__ = {'sqlite_autoincrement': True}
-    course_id = db.Column(db.Integer, primary_key=True)
-    course_number = db.Column(db.Integer, nullable=False)
-    course_name = db.Column(db.String(10), nullable=False)
-    year = db.Column(db.Date, nullable=False)
-    term = db.Column(db.String(50), nullable=False)
-    active = db.Column(db.Boolean, nullable=False)
-    admin_id = db.Column(db.Integer, ForeignKey("Course.course_id"), nullable=False)
-    
 def get_courses():
     try:
         return Course.query.all()
