+++ BackEndFlask/models/oc.py
@@ -1,19 +1,10 @@
 from core import db
-from flask_login import UserMixin
-from sqlalchemy import ForeignKey
 from sqlalchemy.exc import SQLAlchemyError
+from models.schemas import ObservableCharacteristics
 
 class InvalidOCID(Exception):
     "Raised when oc_id does not exist!!!"
     pass
-
-class ObservableCharacteristics(UserMixin, db.Model):
-    __tablename__ = "ObservableCharacteristics"
-    __table_args__ = {'sqlite_autoincrement': True}
-    oc_id = db.Column(db.Integer, primary_key=True)
-    rubric_id = db.Column(db.Integer, ForeignKey("Rubric.rubric_id"), nullable=False)
-    category_id = db.Column(db.Integer,ForeignKey("Category.category_id"), nullable=False)
-    oc_text = db.Column(db.String(10000), nullable=False)
 
 def get_OCs():
     try:
