+++ BackEndFlask/models/rubric.py
@@ -1,18 +1,10 @@
 from core import db
-from flask_login import UserMixin
-from sqlalchemy import ForeignKey
 from sqlalchemy.exc import SQLAlchemyError
+from models.schemas import Rubric
 
 class InvalidRubricID(Exception):
     "Raised when rubric_id does not exist!!!"
     pass
-
-class Rubric(UserMixin, db.Model):
-    __tablename__ = "Rubric"
-    __table_args__ = {'sqlite_autoincrement': True}
-    rubric_id = db.Column(db.Integer, primary_key=True)
-    rubric_name = db.Column(db.String(100))
-    rubric_desc = db.Column(db.String(100), nullable=True)
 
 def get_rubrics():
     try:
