+++ BackEndFlask/models/user.py
@@ -1,30 +1,11 @@
 from core import db
-from flask_login import UserMixin
-from sqlalchemy import ForeignKey, BOOLEAN
 from werkzeug.security import generate_password_hash
-
-# tables in database; each class match to a table in database
-#   *size of username, project_id, owner, project_name should be consistent in different tables.
-#   *password is encrypted
 from sqlalchemy.exc import SQLAlchemyError
+from models.schemas import Users
 
 class InvalidUserID(Exception):
     "Raised when user_id does not exist!!!"
     pass
-
-class Users(UserMixin, db.Model):
-    __tablename__ = "Users"
-    __table_args__ = {'sqlite_autoincrement': True}
-    user_id = db.Column(db.Integer, primary_key=True)
-    fname = db.Column(db.String(30), nullable=False)
-    lname = db.Column(db.String(30), nullable=False)
-    email = db.Column(db.String(255), unique=True, nullable=False)
-    password = db.Column(db.String(80), nullable=False)
-    role = db.Column(db.String(20), nullable=False)     #role in university; ex. instructor or ta
-    lms_id = db.Column(db.Integer, unique=True, nullable=True)
-    # Need to change consent to a string that can be either yes, no, or nothing!
-    consent = db.Column(db.Boolean, nullable=False)
-    owner_id = db.Column(db.Integer, ForeignKey("Users.user_id"), nullable=False)
 
 def get_users():
     try:
